deleted comments

+++ controllers/claims.controller.js
@@ -5,7 +5,7 @@ const jwt = require("jsonwebtoken");
 //Crea un nuevo claim dentro del tenant especifico que nos llega con el jwt.
 const createNewUserClaim = async function (req, res) {
   const { jwtToken, claim } = req.body;
-  //const retornoValidate = await ClaimService.validateJwt(jwtToken);
+  const retornoValidate = await ClaimService.validateJwt(jwtToken);
   try {
     if (!retornoValidate) {
       return res.status(401).json({ message: "XX - The token was corrupted." });
@@ -29,11 +29,11 @@ const createNewUserClaim = async function (req, res) {
 
 const deleteClaimFromTenant = async function (req, res) {
   const { jwtToken, claim } = req.body;
-  //const retornoValidate = await ClaimService.validateJwt(jwtToken);
+  const retornoValidate = await ClaimService.validateJwt(jwtToken);
   try {
-    //if (!retornoValidate) {
-      //return res.status(401).json({ message: "XX - The token was corrupted." });
-    //}
+    if (!retornoValidate) {
+      return res.status(401).json({ message: "XX - The token was corrupted." });
+    }
     const { tenant } = jwt.decode(jwtToken);
     var result = await ClaimService.deleteExistingClaim(
       tenant,
